Improve Portfolio calculations with Decimal arithmetic in Rust
- Make `get_exchange_rate` and `get_xrate` return Decimal
- Accumulate portfolio money in Decimal, round once per currency
- Return Decimal from `calculate_xrate_to_base`
- Keep the PyO3 f64 contract; convert at the boundary
- Add non-positive-rate guard and round-once tests
- Related to #4165

+++ nautilus_trader/core/nautilus_pyo3.pyi
@@ -83,7 +83,7 @@ def get_exchange_rate(
     price_type: PriceType,
     quotes_bid: dict[str, float],
     quotes_ask: dict[str, float],
-) -> float | None: ...
+) -> Decimal | None: ...
 
 # Logging
